scripts/wta/flask_app/requirements.txt
# Flask Framework
Flask==3.0.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Database
mysql-connector-python==8.2.0
pymysql==1.1.0
sqlalchemy==2.0.23

# JIRA Integration
jira==3.5.2
atlassian-python-api==3.41.3

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Similarity Comparison
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0

# Additional dependencies
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.110.0
